+++ python/ipywidgets/ipywidgets/widgets/widget.py
@@ -399,6 +399,10 @@ class Widget(LoggingHasTraits):
                     'model_module_version': widget._model_module_version,
                     'state': widget.get_state(drop_defaults=drop_defaults),
                 }
+            if 'widget_id' in data:
+                # In this case, we only want 1 widget state
+                id = data['widget_id']
+                full_state = {k: v for k, v in full_state.items() if k == id}
             full_state, buffer_paths, buffers = _remove_buffers(full_state)
             cls._control_comm.send(dict(
                 method='update_states',
@@ -441,7 +445,9 @@ class Widget(LoggingHasTraits):
         state = {}
         if widgets is None:
             widgets = _instances.values()
+        print(widgets)
         for widget in widgets:
+            print(widget.model_id)
             state[widget.model_id] = widget._get_embed_state(drop_defaults=drop_defaults)
         return {'version_major': 2, 'version_minor': 0, 'state': state}
 
@@ -533,6 +539,7 @@ class Widget(LoggingHasTraits):
                 args['comm_id'] = self._model_id
 
             self.comm = comm.create_comm(**args)
+            print(self.comm)
 
     @observe('comm')
     def _comm_changed(self, change):
